LayoutTests/userscripts/window-onerror-for-isolated-world-1-expected.txt

Test that window.onerror and "error" event listeners from main world are invoked for uncaught exceptions in user scripts running in isolate worlds as well as for exceptions in the main world.Bug 8519.

Main world window.onerror: Error: Error in main world inline script. at window-onerror-for-isolated-world-1.html:54
Main world error event listener: Error: Error in main world inline script. at window-onerror-for-isolated-world-1.html:54
Main world window.onerror: Error: Error in user script inline script. at undefined:11
Main world error event listener: Error: Error in user script inline script. at undefined:11
Main world window.onerror: Error: Error in main world load handler. at window-onerror-for-isolated-world-1.html:46
Main world error event listener: Error: Error in main world load handler. at window-onerror-for-isolated-world-1.html:46
Main world window.onerror: Error: Error in user script load handler. at undefined:3
Main world error event listener: Error: Error in user script load handler. at undefined:3
Main world window.onerror: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-1.html:51
Main world error event listener: Error: Error in main world setTimeout callback. at window-onerror-for-isolated-world-1.html:51
Main world window.onerror: Error: Error in user script setTimeout callback. at undefined:8
Main world error event listener: Error: Error in user script setTimeout callback. at undefined:8
